+++ tests/extension/test_service.py
@@ -2,7 +2,7 @@ from contextlib import contextmanager
 
 import pytest
 
-from antidote import Service, Wiring, service, world
+from antidote import Service, Wiring, service, world, Provide
 from antidote._providers import ServiceProvider
 from antidote.exceptions import DuplicateDependencyError
 
@@ -85,6 +85,55 @@ def test_custom_scope():
     assert world.get(Scoped) is not my_service
 
 
+def test_parameterized():
+    x = object()
+
+    class A(Service):
+        __antidote__ = Service.Conf(parameters=['x'])
+
+        def __init__(self, **kwargs):
+            self.kwargs = kwargs
+
+    a = world.get(A.parameterized(x=x))
+    assert a.kwargs == dict(x=x)
+
+    with pytest.raises(ValueError, match=".*parameters.*'x'.*"):
+        A.parameterized()
+
+    with pytest.raises(ValueError, match=".*parameters.*'x'.*"):
+        A.parameterized(unknown='something')
+
+
+def test_not_parametrized():
+    class A(Service):
+        def __init__(self, **kwargs):
+            self.kwargs = kwargs
+
+    with pytest.raises(RuntimeError, match=".*parameters.*"):
+        A.parameterized(x=1)
+
+
+def test_invalid_with_default_parameters():
+    with pytest.raises(ValueError, match=".*default.*"):
+        class A(Service):
+            __antidote__ = Service.Conf(parameters=['x'])
+
+            def __init__(self, x: str = 'default'):
+                self.x = x
+
+
+def test_invalid_with_injected_parameters():
+    class A(Service):
+        pass
+
+    with pytest.raises(ValueError, match=".*injected.*class.*A.*"):
+        class B(Service):
+            __antidote__ = Service.Conf(parameters=['x'])
+
+            def __init__(self, x: Provide[A]):
+                self.x = x
+
+
 def test_duplicate_registration():
     with pytest.raises(DuplicateDependencyError):
         @service
@@ -129,6 +178,20 @@ def test_invalid_conf():
             __antidote__ = object()
 
 
+@pytest.mark.parametrize('expectation, parameters', [
+    (pytest.raises(TypeError), "string"),
+    (pytest.raises(TypeError), object()),
+    (pytest.raises(TypeError), [1]),
+    (does_not_raise(), ['x']),
+    (does_not_raise(), []),
+    (does_not_raise(), None),
+])
+def test_conf_parameters(expectation, parameters):
+    with expectation:
+        class A(Service):
+            __antidote__ = Service.Conf(parameters=parameters)
+
+
 @pytest.mark.parametrize('expectation, kwargs', [
     pytest.param(pytest.raises(TypeError, match=f'.*{arg}.*'),
                  {arg: object()},
@@ -136,7 +199,7 @@ def test_invalid_conf():
     for arg in ['wiring',
                 'singleton',
                 'scope',
-                'tags']
+                'parameters']
 ])
 def test_invalid_conf_args(kwargs, expectation):
     with expectation:
@@ -145,7 +208,9 @@ def test_invalid_conf_args(kwargs, expectation):
 
 @pytest.mark.parametrize('kwargs', [
     dict(singleton=False),
+    dict(scope=None),
     dict(wiring=Wiring(methods=['method'])),
+    dict(parameters=frozenset(['x']))
 ])
 def test_conf_copy(kwargs):
     conf = Service.Conf(singleton=True).copy(**kwargs)